requirement.txt
asgiref==3.2.9
astroid==2.4.2
Django==3.0.7
django-tastypie==0.14.3
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
